updated migration workspace for to use tinyInteger  and not Integer for boolean

+++ database/migrations/2023_08_28_002311_create_workspaces_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('workspaces', function (Blueprint $table) {
             $table->id();
             $table->string('name')->nullable();
-            $table->integer('config');
+            $table->tinyInteger('config');
 
             $table->timestamps();
         });
